+++ src/test/java/com/stratio/qa/specs/KafkaSecStepsIT.java
@@ -22,8 +22,8 @@ import com.stratio.qa.utils.BaseGTest;
 import cucumber.api.CucumberOptions;
 import org.testng.annotations.Test;
 
-@CucumberOptions(plugin = "json:target/cucumber.json", features = {"src/test/resources/features/kafkaSteps.feature"})
-public class KafkaStepsIT extends BaseGTest {
+@CucumberOptions(plugin = "json:target/cucumber.json", features = {"src/test/resources/features/kafkaSecSteps.feature"})
+public class KafkaSecStepsIT extends BaseGTest {
 
     @Test(dataProvider = "scenarios")
     public void run(PickleEventWrapper pickleWrapper, CucumberFeatureWrapper featureWrapper) throws Throwable {